Moves KeyValuePairs type into its own module
